+++ src/robolli_legacy.cpp
@@ -8,9 +8,9 @@ using namespace walkman::drc::valve;
 /** NOTE this are board ids, not array positions */
 int r_arm[] =       { 16, 17, 18 ,19};
 int l_arm[] =       { 20, 21, 22, 23};
-int r_arm_offs[] =  { 90,  0,  0,  0};  // offset in y2r(x)
-int l_arm_offs[] =  {-90,  0,  0,  0};  // offset in y2r(x)
-int no_offs[] =     {  0,  0,  0,  0};
+double r_arm_offs[] =  { 90,  0,  0,  0};  // offset in y2r(x)
+double l_arm_offs[] =  {-90,  0,  0,  0};  // offset in y2r(x)
+double no_offs[] =     {  0,  0,  0,  0};
 int r_farm[] = { 26, 27, 28, 32};
 int l_farm[] = { 29, 30, 31, 33};
 
@@ -80,8 +80,8 @@ unsigned long int robolli_legacy::get_time_ns() {
 
 void robolli_legacy::getRobolliLeftArm(const int* robolliVec,
                                        double* left_arm_data,
-                                       const unsigned int scale,
-                                       const int* y2r_offset) {
+                                       const double scale,
+                                       const double* y2r_offset) {
     for(unsigned int i = 0; i < l_arm_size; ++i)
         left_arm_data[i] = (robolliVec[l_arm[i]] -  y2r_offset[i])/scale;
     for(unsigned int i = 0; i < l_farm_size; ++i)
@@ -90,8 +90,8 @@ void robolli_legacy::getRobolliLeftArm(const int* robolliVec,
 
 void robolli_legacy::getRobolliRightArm(const int* robolliVec,
                                         double* right_arm_data,
-                                        const unsigned int scale,
-                                        const int* y2r_offset) {
+                                        const double scale,
+                                        const double* y2r_offset) {
     for(unsigned int i = 0; i < r_arm_size; ++i)
         right_arm_data[i] = (robolliVec[r_arm[i]] - y2r_offset[i])/scale;
     for(unsigned int i = 0; i < r_farm_size; ++i)
@@ -100,8 +100,8 @@ void robolli_legacy::getRobolliRightArm(const int* robolliVec,
 
 void robolli_legacy::setRobolliLeftArm(int* robolliVec,
                                        const double* left_arm_data,
-                                       const unsigned int scale,
-                                       const int* y2r_offset) {
+                                       const double scale,
+                                       const double *y2r_offset) {
     for(unsigned int i = 0; i < l_arm_size; ++i)
         robolliVec[l_arm[i]] = (left_arm_data[i] + y2r_offset[i])*scale;
     for(unsigned int i = 0; i < l_farm_size; ++i)
@@ -110,8 +110,8 @@ void robolli_legacy::setRobolliLeftArm(int* robolliVec,
 
 void robolli_legacy::setRobolliRightArm(int* robolliVec,
                                         const double* right_arm_data,
-                                        const unsigned int scale,
-                                        const int* y2r_offset) {
+                                        const double scale,
+                                        const double* y2r_offset) {
     for(unsigned int i = 0; i < r_arm_size; ++i)
         robolliVec[r_arm[i]] = (right_arm_data[i] + y2r_offset[i])*scale;
     for(unsigned int i = 0; i < r_farm_size; ++i)
